requirements.txt
dvc
dvc-s3
dagshub
scikit-learn
mlflow
pandas
pyyaml
numpy